requirements.txt
absl-py>=0.7.1
astor>=0.8.0
attrs>=19.1.0
backcall>=0.1.0
bleach>=3.1.0
colorama>=0.4.1
cycler>=0.10.0
decorator>=4.4.0
defusedxml>=0.6.0
entrypoints>=0.3
et-xmlfile>=1.0.1
gast>=0.2.2
google-pasta>=0.1.7
grpcio>=1.22.0
h5py>=2.9.0
imageio>=2.5.0
imutils>=0.5.2
ipykernel>=5.1.1
ipython>=7.7.0
ipython-genutils>=0.2.0
ipywidgets>=7.5.0
jdcal>=1.4.1
jedi>=0.14.1
Jinja2>=2.10.1
joblib>=0.13.2
jsonschema>=3.0.1
jupyter>=1.0.0
jupyter-client>=5.3.1
jupyter-console>=6.0.0
jupyter-core>=4.5.0
Keras>=2.2.5
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.0
kiwisolver>=1.1.0
Markdown>=3.1.1
MarkupSafe>=1.1.1
matplotlib>=3.1.0
mistune>=0.8.4
nbconvert>=5.5.0
nbformat>=4.4.0
networkx>=2.3
notebook>=6.0.0
numpy>=1.16.4
opencv-python>=4.1.0.25
openpyxl>=3.0.4
pandas>=0.25.1
pandocfilters>=1.4.2
parso>=0.5.1
pickleshare>=0.7.5
Pillow>=6.0.0
prometheus-client>=0.7.1
prompt-toolkit>=2.0.9
protobuf>=3.9.0
Pygments>=2.4.2
pyparsing>=2.4.0
pyrsistent>=0.15.4
python-dateutil>=2.8.0
pytz>=2019.2
PyWavelets>=1.0.3
pywinpty>=0.5.5
PyYAML>=5.1.2
pyzmq>=18.0.2
qtconsole>=4.5.2
scikit-image>=0.15.0
scikit-learn>=0.21.3
scipy>=1.3.0
Send2Trash>=1.5.0
six>=1.12.0
tensorboard>=1.9.0
tensorflow>=2.4.0
termcolor>=1.1.0
terminado>=0.8.2
testpath>=0.4.2
tornado>=6.0.3
traitlets>=4.3.2
wcwidth>=0.1.7
webencodings>=0.5.1
Werkzeug>=0.15.5
widgetsnbextension>=3.5.0
wrapt>=1.11.2
